refactor(script/employee/hr): Moved general employee script to employee module script
